requirements.txt
# Python 3.10.8
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.6
ml-dtypes==0.2.0
mysql-connector-python==9.4.0
numpy==1.26.0
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.15.3
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
